requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
django-environ==0.4.5
google-cloud-secret-manager==2.6.0
django-storages[google]==1.11.1
